Import version 86.09
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@32 8aca7d54-2c30-db11-9de9-000461428c89

+++ simconvoi.h
@@ -645,7 +645,7 @@ public:
      * Removes vehicles at position i
      * @author Hj. Malthaner
      */
-    vehikel_t * remove_vehikel_bei(int i);
+    vehikel_t * remove_vehikel_bei(unsigned short i);
 
 
     /**
